chatbot-ASL/requirements.txt
flask==3.0.0
flask-cors==4.0.0
requests==2.31.0
beautifulsoup4==4.12.2
pandas==2.1.3
python-dotenv==1.0.0
google-generativeai==0.3.2
gunicorn==21.2.0
streamlit==1.29.0